requirements.txt
psycopg==3.1.13
psycopg2-binary==2.9.9
typing_extensions==4.8.0
